added town data to teams
